ABS.Runtime

Documentation

data Null'

Uninhabited. It is used to initially-type the null static object and the this in (ABS' this) of main block.

type ABS' = ContT () IO

Our ABS stateful code is actually an IO-enabled semi-coroutine.

type IORef' = IORef

Local variables in stateful code are actually GHC's heap mutable references.

Note: this type synonym is redundant. It is only used for haddock documentation and not actual code-generation.

type Fut a = MVar a

A future reference is a write-once locking var.

Write-once is not imposed by Haskell, but we assume our translation respects this

NB: we deviate from ABS by not providing ordering of future-refs

data Cog

A COG reference is a mutex and a sleep table.

A process becomes active by acquiring&holding the COG's lock. A process deactivates (by suspend,await) by releasing the lock.

Constructors

Cog (IORef SleepTable) (TQueue (ABS' ()))

type SleepTable = [(IO Bool, ABS' ())]

The sleep table is a naive assoc-list of boolean conditions to their suspend continuations.

data Obj' contents

an object reference is a pair of 1) a reference to its cog 2) its attributes placed in a mutable variable=IORef

NB: we deviate from ABS by not providing ordering of object-refs

Constructors

Obj' (IORef contents) !Cog

class Sub' sub sup where

Subtyping-relation for ABS objects (as a multiparam typeclass).

Minimal complete definition

up'

Methods

up' :: sub -> sup

The upcasting method from a subtype to a supertype

apiStore' :: IORef (Map String Dynamic)

A dynamic key-value store that stores the HTTPCallable names of objects to their heap reference (IORef).

This technique effectively disables the garbage-collection for any of the exposed objects, since the reference is stored as a Dynamic and not as a weak pointer.

Creating new objects

new :: (Obj' a -> IO ()) -> a -> IO (Obj' a)

New object inside a newly-creted COG

unlifted to ABS' monad

newlocal' :: Obj' this -> (Obj' a -> IO ()) -> a -> IO (Obj' a)

New object inside the currently-executing COG

unlifted to ABS' monad

Synchronous method calling

sync' :: Obj' this -> Obj' a -> (Obj' a -> ABS' r) -> ABS' r

sync

(<..>) :: Obj' a -> (Obj' a -> ABS' r) -> ABS' r

optimized sync, by not running same-cog-check. Used only by the generation when stumbled on "this.m()".

Asynchronous method calling

(<!>) :: Obj' a -> (Obj' a -> ABS' b) -> IO (Fut b)

async, unlifted

(<!!>) :: Obj' a -> (Obj' a -> ABS' b) -> IO ()

fire&forget async, unlifted

awaitSugar'

Arguments

:: Obj' this

-> (b -> IO ())
LHS

-> Obj' a
callee

-> (Obj' a -> ABS' b)
method

-> ABS' ()

Specialized (optimized support) for A res = await o!m().

The implementors could instead use a non-specialized version which desugars to FutA f_fresh = o!m(); await f_fresh; A res = f_fresh.get;

awaitSugar''

Arguments

:: Obj' this

-> Obj' a
callee

-> (Obj' a -> ABS' b)
method

-> ABS' ()

Similar to awaitSugar' but does not have a left-hand side, i.e. is of the form await o!m();.

Cooperative yielding control

suspend :: Obj' this -> ABS' ()

Optimized suspend by avoiding capturing current-continuation if the method will be reactivated immediately: implemented by inlining back' function and using a custom TQueue that exposes the otherwise abstract datatype TQueue.

Awaiting on boolean conditions

awaitBool' :: Obj' thisContents -> (thisContents -> Bool) -> ABS' ()

Awaiting on futures

awaitFuture' :: Obj' this -> Fut a -> ABS' ()

awaitFutures' :: Obj' this -> [IO Bool] -> IO a -> ABS' ()

awaitFutureField' :: Typeable a => Obj' this -> (([ThreadId] -> [ThreadId]) -> this -> this) -> Fut a -> ABS' ()

data ChangedFuture' a

Constructors

ChangedFuture' (Fut a)

Blocking on futures

get :: Fut b -> IO b

Block until the future has been resolved and try to extract the value. May re-raise the exception of the original failed future. unlifted to ABS' monad

Primitives for Timed-ABS extension

(using a soft-realtime interpretation)

currentms :: IO (Ratio Int)

Not much used. Baggage from the standard ABS library (abslang.abs)

now :: IO TimeSpec

Returns the current time of the real-time clock.

duration :: Ratio Int -> Ratio Int -> IO ()

in seconds, ignores second argument tmax

awaitDuration' :: Obj' this -> Ratio Int -> Ratio Int -> ABS' ()

in seconds, ignores second argument tmax

Lifting ABS pure code to ABS object layer

Haskell is pure by default. These are necessary functions for lifting pure ABS expressions (of the functional core) to the ABS' object layer (monadic statements).

Haskell's return is an expression taking a pure value and lifting it to the monadic world.

return :: Monad m => forall a. a -> m a

Inject a value into the monadic type.

ABS strictness

Haskell is lazy be default. We use the HNF-strict version of fmap (<$>) to strictly pass the arguments to ABS methods and statements

(<$!>) :: Monad m => (a -> b) -> m a -> m b infixl 4

Strict version of <$>.

Since: 4.8.0.0

Other

skip :: ABS' ()

Only as a showcase. The translator does not translate but strips away skip statements.

while :: IO Bool -> ABS' () -> ABS' ()

The ABS controlow flow. Takes a mutable boolean condition that may contain fields (thus inside IO) and an ABS block to execute (repeatedly).

while' :: IO Bool -> IO () -> IO ()

Specialized version of while that is used only inside init blocks.

TODO: This version would be unnecessary if the while was abstracted over any (MonadIO a)

random :: Int -> IO Int

Note: this is multicore-safe (thread-safe) but not multicore-scalable, because it uses underneath atomicModifyIORef

main_is' :: (Obj' contents -> ABS' ()) -> ScottyM () -> IO ()

This function takes an ABS'' main function in the module and executes the ABS' program.

Note the mainABS' function expects a this object as input. This is only for unifying the method-block generation; the code-generator will safely catch if a main contains calls to this. This runtime, however, does not do such checks; if the user passes a main that uses this, the program will err.

null :: Obj' Null'

A static heap representation of the null object.

nullFuture' :: Fut a

The uninitialized future.

Calling FutA f; await f? will deadlock the current process. Calling FutA f; f.get; will deadlock the whole current-COG.

I/O primitives of ABS to the terminal screen

print :: String -> IO ()

Decided to be a statement and not a built-in function for keeping functions pure.

println :: String -> IO ()

Decided to be a statement and not a built-in function for keeping functions pure.

readln :: IO String

Decided to be a statement and not a built-in function for keeping functions pure.

ABS extension for exception support

throw :: (MonadThrow m, Exception e) => e -> m a

Throws the given exception. Can be used unlike Haskell only in ABS stateful code.

catch :: MonadCatch m => m a -> [Handler' m a] -> m a

Catches different sorts of exceptions. See Control.Exception's catches

data Handler' m a

different handler because we want maybe (for ABS exception pattern matching)

Constructors

Exception e => Handler' (e -> Maybe (m a))

finally :: MonadCatch m => m () -> m () -> m ()

Our ABS finally is different than default Haskell's, since we do not have async. exceptions, thus we do not need to mask

Same exceptions as Haskell's, so import them rather than redefining. This has the benefit of catching exceptions from the FFI

newtype PatternMatchFail :: *

A pattern match failed. The String gives information about the source location of the pattern.

Constructors

PatternMatchFail String

newtype RecSelError :: *

A record selector was applied to a constructor without the appropriate field. This can only happen with a datatype with multiple constructors, where some fields are in one constructor but not another. The String gives information about the source location of the record selector.

Constructors

RecSelError String

data ArithException :: *

Arithmetic exceptions.

Constructors

DivideByZero

newtype AssertionFailed :: *

assert was applied to False.

Constructors

AssertionFailed String

the ABS-specific builtin exceptions

data NullException

Constructors

NullException

data ABSException'

Constructors

Exception e => ABSException' e

absExceptionToException' :: Exception e => e -> SomeException

absExceptionFromException' :: Exception e => SomeException -> Maybe e

ABS extension to support write-once promises

pro_new :: IO (Fut a)

Newly-allocated promise with empty contents

pro_try :: Fut b -> IO (Maybe b)

Try to extract the value inside the promise.

Does not block.

pro_give :: Fut b -> b -> IO ()

Writes (once) the value to the given promise, essentially resolving it.

Will throw a PromiseRewriteException when to write more than once.

data PromiseRewriteException

Trying to write to an already-resolved promise

Constructors

PromiseRewriteException